able to exclude bits of n-ui
