--- lib/transforms/external-n-ui.js
@@ -1,14 +1,0 @@
-const Wrap = require('../addons/wrap');
-
-module.exports = function (options, output) {
-	if (options.externals && options.externals['n-ui']) {
-
-		output.plugins.push(
-			new Wrap(
-				'(function(){function init(){\n',
-				'\n};window.ftNextnUiLoaded ? init() : document.addEventListener ? document.addEventListener(\'ftNextnUiLoaded\', init) : document.attachEvent(\'onftNextnUiLoaded\', init);})();',
-				{ match: /\.js$/ }
-			)
-		);
-	}
-}
